requirements.txt
numpy==1.19.2
tensorflow==2.5.0
Pillow==8.2.0
mediapipe==0.8.6
opencv-python==4.5.2.54
black
isort
flake8